start -f, stop and status python windows (no cygwin involved)

+++ servicemanager/smprocess.py
@@ -1,19 +1,22 @@
 #!/usr/bin/env python
-from signal import SIGINT, SIGKILL
+from signal import SIGINT
 import os
 import re
+import psutil
 
 from servicemanager import subprocess
 
 
 def kill_pid(context, pid, force=False):
 
-    if _is_system_or_smserver_or_test_process(pid):
-        return "Not allowed to kill system, test or smserver process (pid = %d)" % pid
+    if os.name != "nt":
+        if _is_system_or_smserver_or_test_process(pid):
+           return "Not allowed to kill system, test or smserver process (pid = %d)" % pid
 
     try:
         if force:
-            os.kill(pid, SIGKILL)
+            psutil.Process(pid).kill
+
         else:
             os.kill(pid, SIGINT)
 
@@ -143,21 +146,28 @@ class SmProcess:
 
     @classmethod
     def all_processes(cls):
-        command = "ps -eo ppid,pid,etime,rss,args"
-        ps_command = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE)
-        stdout, stderr = ps_command.communicate()
-        ps_output = stdout.split("\n")[:-1]
+        result = []
 
-        def process_line_to_object(process_line):
-            values = process_line.strip().split()
-            return SmProcess(int(values[0]), int(values[1]), values[2], values[3], values[4:])
+        for proc in psutil.process_iter():
+            try:
+                smprocess = SmProcess( proc.ppid(),
+                                       proc.pid,
+                                       proc.create_time(),
+                                       proc.memory_info().rss,
+                                       proc.cmdline() )
+            except psutil.NoSuchProcess:
+                pass
+            else:
+                result.append(smprocess)
 
-        return map(process_line_to_object, ps_output[1:])
+        return result
 
     @staticmethod
     def find_in_command_line(process, r):
         for arg in process.args:
             if r.search(arg):
+                if os.name == "nt" and "cmd.exe" in process.args[0]:
+                    return False
                 return True
         return False
